+++ podaac/lambda_handler/lambda_handler.py
@@ -8,7 +8,6 @@ from shutil import rmtree
 import requests
 
 import botocore
-import numpy as np
 import xarray as xr
 from cumulus_logger import CumulusLogger
 from cumulus_process import Process, s3
@@ -178,28 +177,15 @@ class FootprintGenerator(Process):
             self.logger.error("Error downloading granule from s3: {}".format(ex), exc_info=True)
             raise ex
 
-        with open(config_file) as config_f:
-            read_config = json.load(config_f)
+        strategy, footprint_params = forge.load_footprint_config(config_file)
+        footprint_params["path"] = self.path
+        with xr.open_dataset(local_file, group=footprint_params.get('group'), decode_times=False) as ds:
+            lon_data = ds[footprint_params['longitude_var']]
+            lat_data = ds[footprint_params['latitude_var']]
 
-        longitude_var = read_config.get('lonVar')
-        latitude_var = read_config.get('latVar')
-        is360 = read_config.get('is360', False)
-
-        thinning_fac = read_config.get('footprint', {}).get('thinning_fac', 100)
-        alpha = read_config.get('footprint', {}).get('alpha', 0.05)
-        strategy = read_config.get('footprint', {}).get('strategy', None)
-        simplify = read_config.get('footprint', {}).get('simplify', 0.1)
-        group = read_config.get('footprint', {}).get('group')
-        cutoff_lat = read_config.get('footprint', {}).get('cutoff_lat', None)
-        smooth_poles = read_config.get('footprint', {}).get('smooth_poles', None)
-        fill_value = read_config.get('footprint', {}).get('fill_value', np.nan)
-
-        # Generate footprint
-        with xr.open_dataset(local_file, group=group, decode_times=False) as ds:
-            lon_data = ds[longitude_var]
-            lat_data = ds[latitude_var]
-            wkt_representation = forge.generate_footprint(lon_data, lat_data, thinning_fac=thinning_fac, alpha=alpha, is360=is360, simplify=simplify,
-                                                          cutoff_lat=cutoff_lat, smooth_poles=smooth_poles, strategy=strategy, fill_value=fill_value)
+            wkt_representation = forge.generate_footprint(
+                lon_data, lat_data, strategy=strategy, **footprint_params
+            )
 
         wkt_json = {
             "FOOTPRINT": wkt_representation,
@@ -236,11 +222,6 @@ class FootprintGenerator(Process):
             Payload that is returned to the cma which is a dictionary with list of granules
         """
 
-        config_file_path = self.get_config()
-
-        granules = self.input['granules']
-        append_output = {}
-
         if self.kwargs.get('context', None):
             try:
                 aws_request_id = self.kwargs.get('context').aws_request_id
@@ -252,6 +233,11 @@ class FootprintGenerator(Process):
                 self.logger.info(message)
             except AttributeError:
                 pass
+
+        config_file_path = self.get_config()
+
+        granules = self.input['granules']
+        append_output = {}
 
         for granule in granules:
             granule_id = granule['granuleId']